Adding the Historian
I changed a lot of things around. If we decide that the design changes
I made don’t work, that’s fine. I won’t be offended. I made the changes
I made because I thought it was the best way to clean up the HTML and
add in these new pictures.